Added list view for kids

+++ resources/views/livewire/item-index.blade.php
@@ -3,7 +3,7 @@
         @foreach ($items as $item)
             <div
                 wire:key="{{ $item->id }}"
-                class="relative flex items-center space-x-3 rounded-lg border border-gray-300 bg-white px-6 py-5 shadow-sm focus-within:ring-2 focus-within:ring-indigo-500 focus-within:ring-offset-2 hover:border-gray-400"
+                class="relative flex items-center space-x-3 rounded-lg border border-gray-300 bg-white px-6 py-5 shadow-sm hover:border-gray-400"
             >
                 <div class="min-w-0 flex-1">
                     <div class="focus:outline-none">
@@ -40,14 +40,16 @@
                                         </flux:navmenu.item>
                                     @endif
 
-                                    <flux:navmenu.item
-                                        wire:click="deleteItem({{ $item->id }})"
-                                        wire:confirm="Are you sure you want to delete this item?"
-                                        variant="danger"
-                                        icon="trash"
-                                    >
-                                        Delete
-                                    </flux:navmenu.item>
+                                    @if (! $disableDelete)
+                                        <flux:navmenu.item
+                                            wire:click="deleteItem({{ $item->id }})"
+                                            wire:confirm="Are you sure you want to delete this item?"
+                                            variant="danger"
+                                            icon="trash"
+                                        >
+                                            Delete
+                                        </flux:navmenu.item>
+                                    @endif
                                 </flux:navmenu>
                             </flux:dropdown>
                         </div>
